chore(server-split): v0.4.x PR-G — remaining /admin/* routes → routes/admin.py (47 endpoints + 19 helpers/models)
docs/design/v0.4.x-server-split.md 8-PR sequence 여덟 번째 PR.
PR-F 위에 잔여 /admin/* 도메인 전체 분리 (가장 큰 PR).

## 이동된 surface (2002 lines)

### 47 endpoints (47 /admin/* + 2 helpers)
- /admin/web-search-status, /admin/web-search-config/
- /admin/performance/* × 3
- /admin/character/* × 3
- /admin/knowledge/, /admin/dashboard
- /admin/entities × 2, /admin/wiki/resolve-relations
- /admin/graph/* × 7 (snapshot, events, relation × 3, node, event)
- /admin/memory, /admin/trace/{trace_id}, /admin/episodic/{session_id}
- /admin/metrics, /admin/audit/list
- /admin/features/* × 3
- /admin/files/* × 6 (tree, search, view, download, DELETE, PUT)
- /admin/settings, /admin/persona × 2, /admin/settings/cognitive × 2
- /admin/cr/* × 6 (Change Request workflow)

### Pydantic 모델 × 9
WebSearchConfigUpdate, TraitUpdateRequest, FeatureOverrideRequest, FeatureResetRequest, PersonaRequest, AdminSettingsRequest, CognitiveFlagsRequest, _CrProposeRequest/_CrApproveRequest/_CrRejectRequest/_CrReviewRequest

### Module helpers + constants × 10
web_search_status (impl), _require_graph_edit_enabled, _truncate_audit_blob, _AUDIT_CATEGORIES, _file_mgmt_roots, _FILE_DOWNLOAD_ALLOWED_EXTS, _resolve_under_root, _FILE_VIEW_TEXT_EXTS, _cr_mod/_cr_apply imports, _cr_as_dict, _review_as_dict

## routes/admin.py — _LazySingleton 패턴

47 endpoints 중 다수가 `rag_engine.X` / `file_processor.X` 를 직접
사용. 47개 핸들러마다 shim 추가 대신 모듈-level `_LazySingleton`
descriptor 도입:

```python
class _LazySingleton:
    def __init__(self, getter):
        object.__setattr__(self, "_getter", getter)
    def __getattr__(self, name):
        return getattr(self._getter(), name)

rag_engine = _LazySingleton(get_rag_engine)
file_processor = _LazySingleton(get_file_processor)
```

handler 본문은 byte-identical 유지 (rag_engine.foo 그대로). 모든
attribute access 가 호출 시점 forwarder 통해 routes/_deps 의
get_*() 로 위임 — 테스트 monkeypatch 자연 통과.

## Back-compat re-exports (server_llmwiki.py)

테스트들이 `srv._file_mgmt_roots`, `srv._AUDIT_CATEGORIES` 등으로
attribute 접근. 4개 helper/constant re-import (#noqa: F401).

## tests 갱신 (10 파일)

source-scan 7개 + audit-DB monkeypatch 2개 + graph_snapshot 1개:
- test_admin_entities / metrics / trace / wiki_resolve_relations / file_mgmt_tab / multimodal_trusted / web_search_admin_panel / graph_snapshot: `combined_server_source()` 사용
- test_admin_audit_endpoint / dashboard_audit_sqlite: srv/routes._helpers/routes.admin 3 곳 모두 `_AUDIT_DB` patch
- tests/_server_split_helpers.py: `routes.admin` 추가

## 검증

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ 125 endpoints identical to origin/main |
| 2-5. middleware/audit/static/startup | ✓ smoke pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** |
| ruff F-class | ✓ All checks passed |

server_llmwiki.py: 3537 → 1535 lines (-2002).
8-PR roadmap 진행률 ≈ **74%** (4359/5894).

## Out of scope
- PR-H: server 최종 정리 (target ~500 lines)

Quality delta: exempt (label: chore). URL byte-identical, core/ 무관.
