+++ tests/test_admin_entities.py
@@ -16,7 +16,6 @@ Run:
 """
 from __future__ import annotations
 
-import inspect
 import os
 import sys
 import unittest
@@ -32,8 +31,8 @@ class BackendEndpointContractTests(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        import server_llmwiki as srv
-        cls.src = inspect.getsource(srv)
+        from tests._server_split_helpers import combined_server_source
+        cls.src = combined_server_source()
 
     def test_list_endpoint_accepts_query_filters(self):
         # Find the /admin/entities decorator + handler.
